Always use pip with the flit backend
This patch removes the dependency text files and changes the Makefile to
always use pip as the frontend and flit as the backend.  Note that the
editable install via pip requires flit 3.5 and pip 21.3.  As 21.3.1
fixes a bug related to editable installs, we use it as the minimum
version.  This does not affect the minimum Python version 3.6.

Fixes #111.